Move pure views to views folder
